Create BlackJack
So some of the class names are labeled as go fish and not black jack. This is because we were going to do a go fish game but decided that black jack would be easier and did not want to refactor and rename all of the code to reflect this decision.